refactor(windows-security): hoist sid_to_string into lib.rs; drop noise
Per review feedback: de-duplicate the PSID→SDDL helper so the
follow-up PR's SID-derivation helpers can reuse the same wrapper
instead of inlining their own. Also drop a verbose 'why no async
stop()' comment in register-sparse main.
